@@ -1,29 +1,27 @@
-using System.Collections.Generic;
 using Deceit.Backend.Domain.Game.Players;
 using Deceit.Backend.Domain.Game.SceneCards;
 
-namespace Deceit.Backend.Domain.Game
+namespace Deceit.Backend.Domain.Game;
+
+/// <summary>
+/// The representation of the game state that is
+/// accessible to Investigators
+/// </summary>
+class InvestigatorGameState : PlayerGameState
 {
-    /// <summary>
-    /// The representation of the game state that is
-    /// accessible to Investigators
-    /// </summary>
-    class InvestigatorGameState : PlayerGameState
+    public string PlayerRole { get; }
+
+    public IEnumerable<SceneCard> SceneCards { get; }
+
+    public IEnumerable<InvestigatorWithoutRole> Investigators { get; }
+
+    public InvestigatorGameState(
+        string playerRole,
+        IEnumerable<SceneCard> sceneCards,
+        IEnumerable<InvestigatorWithoutRole> investigators)
     {
-        public string PlayerRole { get; }
-
-        public IEnumerable<SceneCard> SceneCards { get; }
-
-        public IEnumerable<InvestigatorWithoutRole> Investigators { get; }
-
-        public InvestigatorGameState(
-            string playerRole,
-            IEnumerable<SceneCard> sceneCards,
-            IEnumerable<InvestigatorWithoutRole> investigators)
-        {
-            PlayerRole = playerRole;
-            SceneCards = sceneCards;
-            Investigators = investigators;
-        }
+        PlayerRole = playerRole;
+        SceneCards = sceneCards;
+        Investigators = investigators;
     }
 }
