@@ -1,7 +1,6 @@
-namespace Deceit.Backend.Domain.Game
-{
-    public abstract class PlayerGameState
-    {
+namespace Deceit.Backend.Domain.Game;
 
-    }
+public abstract class PlayerGameState
+{
+
 }
