@@ -1,15 +1,14 @@
-namespace Deceit.Backend.Domain.Game.Players
-{
-    /// <summary>
-    /// The Forensic Scientist player
-    /// </summary>
-    class ForensicScientist
-    {
-        public string PlayerId { get; }
+namespace Deceit.Backend.Domain.Game.Players;
 
-        public ForensicScientist(string playerId)
-        {
-            PlayerId = playerId;
-        }
+/// <summary>
+/// The Forensic Scientist player
+/// </summary>
+class ForensicScientist
+{
+    public string PlayerId { get; }
+
+    public ForensicScientist(string playerId)
+    {
+        PlayerId = playerId;
     }
 }
