@@ -1,12 +1,10 @@
-using System.Threading.Tasks;
 using Deceit.Backend.Domain.Game;
 using Deceit.Backend.Domain.Lobbies;
 
-namespace Deceit.Backend.Hubs
+namespace Deceit.Backend.Hubs;
+
+public interface IPreGameHubClient
 {
-    public interface IPreGameHubClient
-    {
-        Task LobbyUpdated(Lobby players);
-        Task StartGame(PlayerGameState playerGameState);
-    }
+    Task LobbyUpdated(Lobby players);
+    Task StartGame(PlayerGameState playerGameState);
 }
