@@ -1,30 +1,25 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Deceit.Backend.Domain.Lobbies;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace Deceit.Backend.Pages
+namespace Deceit.Backend.Pages;
+
+public class IndexModel : PageModel
 {
-    public class IndexModel : PageModel
+    private readonly LobbyService lobbyService;
+
+    public IndexModel(LobbyService lobbyService)
     {
-        private readonly LobbyService lobbyService;
+        this.lobbyService = lobbyService;
+    }
 
-        public IndexModel(LobbyService lobbyService)
-        {
-            this.lobbyService = lobbyService;
-        }
+    public RedirectToPageResult OnPost()
+    {
+        // generate new random lobby
+        string lobbyId = Guid.NewGuid().ToString();
+        lobbyService.AddLobby(new Lobby(lobbyId));
 
-        public RedirectToPageResult OnPost()
-        {
-            // generate new random lobby
-            string lobbyId = Guid.NewGuid().ToString();
-            lobbyService.AddLobby(new Lobby(lobbyId));
-
-            // redirect user to it
-            return RedirectToPage("Lobby", new { lobbyId });
-        }
+        // redirect user to it
+        return RedirectToPage("Lobby", new { lobbyId });
     }
 }
