@@ -1,17 +1,14 @@
-using System.Collections.Generic;
+namespace Deceit.Backend.Domain.Game.SceneCards;
 
-namespace Deceit.Backend.Domain.Game.SceneCards
+class SceneCard
 {
-    class SceneCard
+    public string Title { get; }
+
+    public IEnumerable<string> Options { get; }
+
+    public SceneCard(string title, IEnumerable<string> options)
     {
-        public string Title { get; }
-
-        public IEnumerable<string> Options { get; }
-
-        public SceneCard(string title, IEnumerable<string> options)
-        {
-            Title = title;
-            Options = options;
-        }
+        Title = title;
+        Options = options;
     }
 }
